Switch to daisyUI from Skeleton
Skeleton is a bit heavy for this site. Even with Arch's logo SVG added,
the artifacts.zip for the built site is now about 20% smaller.

A light switch is not needed as prefers-color-scheme now works
out-of-the-box. Might be nice to use theme-change[1] either way, but
getting it to play nicely with prefers-color-scheme requires some work.

[1]: https://github.com/saadeghi/theme-change/blob/1d9932c0ed8b743ea2ebd5941d98e35abd1298f0/README.md#toggle-between-two-themes